# Dense matrix increase performance of map and forEach